backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
ddgs>=9.0.0
httpx==0.28.1
pydantic==2.10.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
redis==5.2.1
python-dotenv==1.0.1
